Split a few modules into definitions and implementation.
This is so the renderer process can import definitions and
not worry about executing any code. In particular, importing
`utils/paths.ts` runs the code that sets up the paths and
this code can't be run in the renderer process.
