+++ pkg/rancher-desktop/config/commandLineOptions.ts
@@ -2,7 +2,8 @@ import { join } from 'path';
 
 import _ from 'lodash';
 
-import { LockedSettingsType, save, Settings, turnFirstRunOff } from '@pkg/config/settings';
+import { LockedSettingsType, Settings } from '@pkg/config/settings';
+import { save, turnFirstRunOff } from '@pkg/config/settingsImpl';
 import { TransientSettings } from '@pkg/config/transientSettings';
 import SettingsValidator from '@pkg/main/commandServer/settingsValidator';
 import Logging from '@pkg/utils/logging';
