+++ pkg/rancher-desktop/window/constants.ts
@@ -1,0 +1,8 @@
+export const mainRoutes = [
+  { route: '/General' },
+  { route: '/PortForwarding' },
+  { route: '/Images' },
+  { route: '/Troubleshooting' },
+  { route: '/Diagnostics' },
+  { route: '/Extensions' },
+];
